fix(update): keep legacy config repair doctor-owned
